vjs_foodiee_app/requirements.txt
Flask==3.0.3
firebase-admin==6.5.0
python-dotenv==1.0.1
gunicorn